requirements.txt
torch
torchvision
scikit-learn
numpy
pandas
matplotlib
seaborn
Pillow
pytorch-grad-cam
opencv-python
scipy
